requirements.txt
fastapi
uvicorn
pytest
pytest-sugar
psycopg2-binary
ruff
strawberry-graphql[fastapi]
strawberry-sqlalchemy-mapper
sqlalchemy
alembic
typer
bs4
requests
arrow
python-dotenv
lxml
# starlette-graphene3
# graphene 
# graphene-sqlalchemy 
pydantic